+++ app/s3push.py
@@ -16,14 +16,15 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
-import sys, os, yaml, hashlib, fnmatch
+import sys, os, yaml, hashlib, fnmatch, logging
 from boto.s3.connection import S3Connection
 
 """ S3Push WebChomp Action Class """
 class webchomp_action:
 
-    def __init__(self, site):
+    def __init__(self, site, config = {}):
         self.site = site
+        self.config = config
 
     def get_webchomp_actions(self):
         return {
@@ -48,7 +49,13 @@ class webchomp_action:
 """ Amazon S3 Uploader Class """
 class webchomp_s3push:
 
-    def __init__(self, site):
+    def __init__(self, site, config = {}):
+
+        # webchomp configuration
+        self.config = config
+
+        # get logger
+        self.logger = logging.getLogger('webchomp')
 
         # set site
         self.site = site
@@ -57,11 +64,12 @@ class webchomp_s3push:
 
         # verify site exists
         if not os.path.exists(self.site_path):
-            raise Exception("Specified site does not exist.")
+            self.logger.critical("Site '%s' does not exist" % site)
+            raise Exception("Site '%s' does not exist" % site)
             return
 
         # Found, yay!
-        print "Found site '%s'..." % (site)            
+        self.logger.info("Found site '%s'" % (site))
 
         # load site conf yml
         self.site_conf = {}
@@ -72,6 +80,7 @@ class webchomp_s3push:
 
         # get s3 key/secret
         if not "amazon_s3" in self.site_conf or not "access_key" in self.site_conf['amazon_s3'] or not 'access_secret' in self.site_conf['amazon_s3']:
+            self.logger.critical("Amazon S3 key/secret have not been set in site.yml.")
             raise Exception("Amazon S3 key/secret have not been set in site.yml.")
             return
 
@@ -80,7 +89,8 @@ class webchomp_s3push:
 
         # get bucket
         if not "bucket" in self.site_conf['amazon_s3']:
-            raise Exception("ERROR: Amazon S3 bucket have not been set in site.yml.")
+            self.logger.critical("Amazon S3 bucket has not been set in site.yml.")
+            raise Exception("Amazon S3 bucket has not been set in site.yml.")
             return
 
         self.s3_bucket_name = self.site_conf['amazon_s3']['bucket']
@@ -90,10 +100,11 @@ class webchomp_s3push:
 
         # get bucket
         if self.s3_conn.lookup(self.s3_bucket_name) is None:
+            self.logger.critical("Bucket '%s' does not exist." % self.s3_bucket_name)
             raise Exception("Bucket '%s' does not exist." % self.s3_bucket_name)
 
         self.s3_bucket = self.s3_conn.get_bucket(self.s3_bucket_name, validate=False)
-        print "Found bucket '%s'..." % self.s3_bucket_name
+        self.logger.info("Found bucket '%s'" % self.s3_bucket_name)
 
         # get list of keys in bucket
         self.s3_bucket_list = {}
@@ -103,30 +114,22 @@ class webchomp_s3push:
     """ Sync local to remote. """
     def sync(self, delete_remote = True):
 
-        print "Begin S3 Sync..."
-
         # check remote files, see if exist locally, delete otherwise
         if delete_remote:
-            print "Performing S3 cleanup...",
             for i in self.s3_bucket_list:
                 if not os.path.exists("%s/%s" % (self.site_output_path, i)):
                     self.s3_bucket_list[i].delete()
-            print "Done"
 
         # get all local files and upload
         for root, dirnames, filenames in os.walk(self.site_output_path):
             for filename in fnmatch.filter(filenames, '*'):
                 self.upload_file( ("%s/%s" % (root, filename)).replace(self.site_output_path, "") ) 
 
-        print "Done!"
-
     """ Upload file to S3 """
     def upload_file(self, file_path):
 
         file_path = os.path.normpath(file_path).replace("\\", "/")
         if file_path[0] == "/": file_path = file_path[1:]
-
-        print "Upload file '%s'..." % file_path,
 
         # check if file exists locally
         if not os.path.exists("%s/%s" % (self.site_output_path, file_path)):
@@ -146,8 +149,10 @@ class webchomp_s3push:
         # compare md5
         # TODO find better way
         if key.etag and key.etag.replace('"',"") == local_md5:
-            print "Skipped (Not changed)"
+            self.logger.info("Upload file: '%s' (Skipped, no change)" % file_path)
             return False
+
+        self.logger.info("Upload file: '%s'" % file_path)
 
         # sudo upload
         key.set_contents_from_filename("%s/%s" % (self.site_output_path, file_path))
@@ -155,5 +160,4 @@ class webchomp_s3push:
         # make public
         key.make_public()
 
-        print "Done"
         return True